cameras and profile refactor
