+++ app/(tabs)/cameras.tsx
@@ -1,42 +1,28 @@
-import { View, StyleSheet, ScrollView, Text, ActivityIndicator, TouchableOpacity, Dimensions } from 'react-native';
-import { WebView } from 'react-native-webview';
-import { useState } from 'react';
-import { Camera, ChevronRight, Shield, AlertCircle, Video, Square } from 'lucide-react-native';
+import React, { useState } from 'react';
+import { ScrollView } from 'react-native';
 
-const { width } = Dimensions.get('window');
+// Импорт компонентов
+import { CameraHeader } from '../components/cameras/CameraHeader';
+import { CameraItemComponent } from '../components/cameras/CameraItem';
+import { CameraFooter } from '../components/cameras/CameraFooter';
 
-const PRIMARY_COLOR = '#8B1E3F'; // бургунди цвет
+// Импорт стилей и данных
+import { styles } from '../components/cameras/CameraStyles';
+import { defaultCameras } from '../components/cameras/CameraData';
 
 export default function CamerasScreen() {
   const [loading, setLoading] = useState(true);
   const [selectedCamera, setSelectedCamera] = useState<number | null>(null);
   const [recordingCameras, setRecordingCameras] = useState<number[]>([]);
-  const [cameras, setCameras] = useState([
-    { 
-      id: 1, 
-      url: 'http://91.210.87.140:8082/mjpg/video.mjpg', 
-      title: 'Подъезд №1',
-      location: 'Вход в подъезд'
-    },
-    { 
-      id: 2, 
-      url: 'http://78.36.19.87/mjpg/video.mjpg', 
-      title: 'Парковка',
-      location: 'Внутренний двор'
-    },
-    { 
-      id: 3, 
-      url: 'http://91.210.87.140:8081/mjpg/video.mjpg', 
-      title: 'Детская площадка',
-      location: 'Внутренний двор'
-    },
-  ]);
+  const [cameras, setCameras] = useState(defaultCameras);
 
   const handleCameraPress = (cameraId: number) => {
     if (selectedCamera === cameraId) {
       setSelectedCamera(null);
     } else {
       setSelectedCamera(cameraId);
+      // Сбрасываем состояние загрузки при каждом переключении камеры
+      setLoading(true);
     }
   };
 
@@ -52,292 +38,27 @@ export default function CamerasScreen() {
     });
   };
 
-  const htmlContent = (url: string) => `
-    <!DOCTYPE html>
-    <html>
-      <head>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <style>
-          body { 
-            margin: 0; 
-            padding: 0; 
-            background: #000;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-          }
-          img { 
-            width: 100%; 
-            height: 100%; 
-            object-fit: cover;
-            border-radius: 12px;
-          }
-          .overlay {
-            position: absolute;
-            top: 0;
-            left: 0;
-            right: 0;
-            bottom: 0;
-            background: linear-gradient(to bottom, rgba(0,0,0,0.2) 0%, rgba(0,0,0,0.6) 100%);
-            border-radius: 12px;
-          }
-        </style>
-      </head>
-      <body>
-        <div style="position: relative; width: 100%; height: 100%;">
-          <img src="${url}" />
-          <div class="overlay"></div>
-        </div>
-      </body>
-    </html>
-  `;
-
+  const handleLoadEnd = () => {
+    setLoading(false);
+  };
 
   return (
     <ScrollView style={styles.container}>
-      <View style={styles.header}>
-        <View style={styles.headerContent}>
-          <Shield size={24} color="#FFFFFF" />
-          <Text style={styles.headerTitle}>Система видеонаблюдения</Text>
-        </View>
-        <Text style={styles.headerSubtitle}>Всего камер: {cameras.length}</Text>
-      </View>
+      <CameraHeader camerasCount={cameras.length} />
 
       {cameras.map((camera) => (
-        <TouchableOpacity 
-          key={camera.id} 
-          style={[
-            styles.cameraContainer,
-            selectedCamera === camera.id && styles.selectedCamera
-          ]}
-          onPress={() => handleCameraPress(camera.id)}
-        >
-          <View style={styles.cameraHeader}>
-            <View style={styles.cameraTitleContainer}>
-              <Text style={styles.cameraTitle}>{camera.title}</Text>
-            </View>
-            <View style={styles.headerControls}>
-              {selectedCamera === camera.id && (
-                <TouchableOpacity 
-                  style={[
-                    styles.recordButton,
-                    recordingCameras.includes(camera.id) && styles.recordingButtonActive
-                  ]}
-                  onPress={(e) => {
-                    e.stopPropagation();
-                    handleRecordPress(camera.id);
-                  }}
-                >
-                  {recordingCameras.includes(camera.id) ? (
-                    <Square size={16} color="#FFFFFF" />
-                  ) : (
-                    <Video size={16} color="#FFFFFF" />
-                  )}
-                </TouchableOpacity>
-              )}
-              <ChevronRight 
-                size={20} 
-                color="#8E8E93"
-                style={[
-                  styles.chevron,
-                  selectedCamera === camera.id && styles.chevronRotated
-                ]}
-              />
-            </View>
-          </View>
-
-          <View style={styles.cameraInfo}>
-            <View style={styles.infoRow}>
-              <Text style={styles.infoLabel}>Расположение:</Text>
-              <Text style={styles.infoValue}>{camera.location}</Text>
-            </View>
-            {recordingCameras.includes(camera.id) && (
-              <View style={styles.recordingIndicator}>
-                <View style={styles.recordingDot} />
-                <Text style={styles.recordingText}>Запись...</Text>
-              </View>
-            )}
-          </View>
-
-          {selectedCamera === camera.id && (
-            <View style={styles.cameraWrapper}>
-              {loading && <ActivityIndicator size="large" color="#007AFF" style={styles.loader} />}
-              <WebView
-                source={{ html: htmlContent(camera.url) }}
-                style={styles.camera}
-                javaScriptEnabled={true}
-                domStorageEnabled={true}
-                startInLoadingState={true}
-                scalesPageToFit={true}
-                onLoadEnd={() => setLoading(false)}
-                onError={(syntheticEvent) => {
-                  const { nativeEvent } = syntheticEvent;
-                  console.warn('WebView error: ', nativeEvent);
-                  setLoading(false);
-                }}
-              />
-            </View>
-          )}
-        </TouchableOpacity>
+        <CameraItemComponent
+          key={camera.id}
+          camera={camera}
+          isSelected={selectedCamera === camera.id}
+          isRecording={recordingCameras.includes(camera.id)}
+          loading={loading && selectedCamera === camera.id}
+          onCameraPress={handleCameraPress}
+          onRecordPress={handleRecordPress}
+        />
       ))}
 
-      <View style={styles.footer}>
-        <AlertCircle size={16} color="#8E8E93" />
-        <Text style={styles.footerText}>
-          Нажмите на камеру для просмотра видеопотока
-        </Text>
-      </View>
+      <CameraFooter />
     </ScrollView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#FFFFFF',
-  },
-  header: {
-    backgroundColor: PRIMARY_COLOR,
-    padding: 20,
-    paddingTop: 60,
-  },
-  headerContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 18,
-    gap: 12,
-    justifyContent: 'center',
-  },
-  headerTitle: {
-    fontSize: 24,
-    fontWeight: '600',
-    color: '#FFFFFF',
-  },
-  headerSubtitle: {
-    fontSize: 14,
-    color: '#FFFFFF',
-    opacity: 0.8,
-    textAlign: 'center',
-  },
-  cameraContainer: {
-    backgroundColor: '#FFFFFF',
-    margin: 16,
-    borderRadius: 12,
-    overflow: 'hidden',
-    shadowColor: '#5a2a37',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  selectedCamera: {
-    borderWidth: 2,
-    borderColor: PRIMARY_COLOR,
-  },
-  cameraHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: 16,
-    borderBottomWidth: 1,
-    borderBottomColor: '#E5E5EA',
-  },
-  cameraTitleContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  headerControls: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 12,
-  },
-  cameraTitle: {
-    fontSize: 16,
-    fontWeight: '600',
-    color: PRIMARY_COLOR,
-    marginBottom: 8,
-  },
-  chevron: {
-    transform: [{ rotate: '0deg' }],
-  },
-  chevronRotated: {
-    transform: [{ rotate: '90deg' }],
-  },
-  cameraInfo: {
-    padding: 16,
-  },
-  infoRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: 8,
-  },
-  infoLabel: {
-    fontSize: 14,
-    color: '#8E8E93',
-  },
-  infoValue: {
-    fontSize: 14,
-    fontWeight: '500',
-  },
-  cameraWrapper: {
-    height: 300,
-    backgroundColor: '#000000',
-    position: 'relative',
-  },
-  camera: {
-    flex: 1,
-  },
-  loader: {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: [{ translateX: -20 }, { translateY: -20 }],
-    zIndex: 1,
-  },
-  footer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 16,
-    backgroundColor: '#FFFFFF',
-    margin: 16,
-    borderRadius: 12,
-  },
-  footerText: {
-    fontSize: 14,
-    color: '#8E8E93',
-    marginLeft: 8,
-  },
-  recordButton: {
-    backgroundColor: '#FF3B30',
-    width: 32,
-    height: 32,
-    borderRadius: 16,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  recordingButtonActive: {
-    backgroundColor: '#34C759',
-  },
-  recordingIndicator: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: '#FFE5E5',
-    padding: 8,
-    borderRadius: 8,
-    marginTop: 8,
-  },
-  recordingDot: {
-    width: 8,
-    height: 8,
-    borderRadius: 4,
-    backgroundColor: '#FF3B30',
-    marginRight: 8,
-  },
-  recordingText: {
-    fontSize: 14,
-    color: '#FF3B30',
-    fontWeight: '500',
-  },
-});
