+++ app/components/cameras/CameraStyles.ts
@@ -1,0 +1,161 @@
+import { StyleSheet, Dimensions } from 'react-native';
+import { PRIMARY_COLOR } from './CameraData';
+
+const { width } = Dimensions.get('window');
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#FFFFFF',
+  },
+  header: {
+    backgroundColor: PRIMARY_COLOR,
+    padding: 20,
+    paddingTop: 60,
+  },
+  headerContent: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 18,
+    gap: 12,
+    justifyContent: 'center',
+  },
+  headerTitle: {
+    fontSize: 24,
+    fontWeight: '600',
+    color: '#FFFFFF',
+  },
+  headerSubtitle: {
+    fontSize: 14,
+    color: '#FFFFFF',
+    opacity: 0.8,
+    textAlign: 'center',
+  },
+  cameraContainer: {
+    backgroundColor: '#FFFFFF',
+    margin: 16,
+    borderRadius: 12,
+    overflow: 'hidden',
+    shadowColor: '#5a2a37',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  selectedCamera: {
+    borderWidth: 2,
+    borderColor: PRIMARY_COLOR,
+  },
+  cameraHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    padding: 16,
+    borderBottomWidth: 1,
+    borderBottomColor: '#E5E5EA',
+  },
+  cameraTitleContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  headerControls: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  cameraTitle: {
+    fontSize: 18,
+    fontWeight: '600',
+    color: '#1A1A1A',
+  },
+  cameraInfoSelected: {
+    padding: 16,
+    borderBottomWidth: 1,
+    borderBottomColor: '#E5E5EA',
+  },
+  cameraInfoNormal: {
+    padding: 16,
+    borderBottomWidth: 0,
+  },
+  infoRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 8,
+  },
+  infoLabel: {
+    fontSize: 14,
+    color: '#8E8E93',
+    marginRight: 8,
+  },
+  infoValue: {
+    fontSize: 14,
+    color: '#1A1A1A',
+  },
+  recordingIndicator: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginTop: 8,
+  },
+  recordingDot: {
+    width: 8,
+    height: 8,
+    borderRadius: 4,
+    backgroundColor: '#FF3B30',
+    marginRight: 6,
+  },
+  recordingText: {
+    fontSize: 14,
+    color: '#FF3B30',
+  },
+  cameraWrapper: {
+    height: width * 0.6,
+    width: '100%',
+    backgroundColor: '#000',
+    position: 'relative',
+  },
+  camera: {
+    width: '100%',
+    height: '100%',
+    backgroundColor: '#000',
+  },
+  loader: {
+    position: 'absolute',
+    left: 0,
+    right: 0,
+    top: 0,
+    bottom: 0,
+    alignItems: 'center',
+    justifyContent: 'center',
+    zIndex: 1,
+  },
+  chevron: {
+    transform: [{ rotate: '0deg' }],
+    marginLeft: 8,
+  },
+  chevronRotated: {
+    transform: [{ rotate: '90deg' }],
+  },
+  recordButton: {
+    width: 32,
+    height: 32,
+    borderRadius: 16,
+    backgroundColor: PRIMARY_COLOR,
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginRight: 8,
+  },
+  recordingButtonActive: {
+    backgroundColor: '#FF3B30',
+  },
+  footer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    padding: 16,
+    paddingBottom: 32,
+  },
+  footerText: {
+    fontSize: 14,
+    color: '#8E8E93',
+    marginLeft: 6,
+  }
+});
