+++ app/components/profile/ProfileStyles.ts
@@ -1,0 +1,122 @@
+import { StyleSheet } from 'react-native';
+import { PRIMARY_COLOR } from './ProfileData';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#FFFFFF',
+  },
+  header: {
+    backgroundColor: '#FFFFFF',
+    padding: 20,
+    paddingTop: 60,
+    alignItems: 'center',
+  },
+  avatar: {
+    width: 80,
+    height: 80,
+    borderRadius: 40,
+    backgroundColor: `${PRIMARY_COLOR}`,
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginBottom: 16,
+  },
+  avatarText: {
+    fontSize: 24,
+    fontWeight: '600',
+    color: PRIMARY_COLOR,
+  },
+  userAddress: {
+    fontSize: 18,
+    color: 'black',
+    opacity: 0.8,
+  },
+  section: {
+    backgroundColor: '#FFFFFF',
+    padding: 20,
+    marginTop: 16,
+  },
+  sectionTitle: {
+    fontSize: 20,
+    fontWeight: '600',
+    marginBottom: 20,
+    textAlign: 'center',
+    color: '#1A1A1A',
+  },
+  infoItem: {
+    marginBottom: 16,
+    paddingBottom: 16,
+    borderBottomWidth: 1,
+    borderBottomColor: '#E5E5EA',
+  },
+  infoLabel: {
+    fontSize: 14,
+    color: '#8E8E93',
+    marginBottom: 4,
+  },
+  infoValue: {
+    fontSize: 16,
+    color: '#1A1A1A',
+  },
+  menuSection: {
+    backgroundColor: `${PRIMARY_COLOR}05`,
+    padding: 16,
+    marginTop: 16,
+  },
+  menuItem: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: '#FFFFFF',
+    padding: 16,
+    borderRadius: 12,
+    marginBottom: 12,
+    shadowColor: '#5a2a37',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 8,
+    elevation: 3,
+  },
+  menuItemTitle: {
+    flex: 1,
+    fontSize: 16,
+    marginLeft: 12,
+    color: '#1A1A1A',
+  },
+  logoutButton: {
+    margin: 16,
+    backgroundColor: PRIMARY_COLOR,
+    padding: 16,
+    borderRadius: 12,
+    alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'center',
+  },
+  logoutText: {
+    color: '#FFFFFF',
+    fontSize: 16,
+    fontWeight: '600',
+    marginLeft: 8,
+  },
+  editButton: {
+    flexDirection: 'row',
+    backgroundColor: PRIMARY_COLOR,
+    padding: 8,
+    paddingHorizontal: 16,
+    borderRadius: 20,
+    marginTop: 10,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  editButtonText: {
+    color: '#FFFFFF',
+    fontSize: 14,
+    fontWeight: '500',
+    marginLeft: 4,
+  },
+  loadingContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#FFFFFF',
+  }
+});
